Clean Up + Move Drivetrain Example to own Class
